+++ juniper-eager-loading/src/lib.rs
@@ -476,7 +476,7 @@
 //! [Juniper]: https://github.com/graphql-rust/juniper
 //! [juniper-from-schema]: https://github.com/davidpdrsn/juniper-from-schema
 
-#![doc(html_root_url = "https://docs.rs/juniper-eager-loading/0.4.1")]
+#![doc(html_root_url = "https://docs.rs/juniper-eager-loading/0.4.2")]
 #![allow(clippy::single_match, clippy::type_complexity)]
 #![deny(
     missing_docs,